requirements.txt
appdirs==1.4.4
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
comm==0.1.2
cryptography==39.0.2
debugpy==1.6.6
decorator==5.1.1
exceptiongroup==1.1.0
executing==1.2.0
frozendict==2.3.5
greenlet==2.0.2
html5lib==1.1
idna==3.4
iniconfig==2.0.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
lxml==4.9.2
matplotlib-inline==0.1.6
multitasking==0.0.11
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.1.0
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.5.post1
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
webencodings==0.5.1
yfinance==0.2.12
